restructure folder structure
