+++ src/main/java/com/example/fortune_cookies_app/controller/Event.java
@@ -1,4 +1,4 @@
-package com.example.fortune_cookies_app;
+package com.example.fortune_cookies_app.controller;
 
 import java.time.LocalDate;
