+++ src/main/java/com/example/fortune_cookies_app/model/EventDAO.java
@@ -1,7 +1,7 @@
-package com.example.fortune_cookies_app.DB;
+package com.example.fortune_cookies_app.model;
 
-import com.example.fortune_cookies_app.Event;
-import com.example.fortune_cookies_app.User;
+import com.example.fortune_cookies_app.controller.Event;
+import com.example.fortune_cookies_app.controller.User;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
