+++ src/main/java/com/example/fortune_cookies_app/model/IUserDAO.java
@@ -1,6 +1,6 @@
-package com.example.fortune_cookies_app.DB;
+package com.example.fortune_cookies_app.model;
 
-import com.example.fortune_cookies_app.User;
+import com.example.fortune_cookies_app.controller.User;
 
 public interface IUserDAO {
     public void createUser(User user);
