+++ src/main/resources/com/example/fortune_cookies_app/add-event-view.fxml
@@ -5,7 +5,7 @@
 <?import javafx.geometry.Insets?>
 
 <VBox xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.fortune_cookies_app.AddEventPaneController"
+      fx:controller="com.example.fortune_cookies_app.controller.AddEventPaneController"
       spacing="10" alignment="TOP_LEFT">
     <padding>
         <Insets top="20" right="20" bottom="20" left="20"/>
