+++ src/test/java/UserTest.java
@@ -1,5 +1,5 @@
 import static org.junit.jupiter.api.Assertions.*;
-import com.example.fortune_cookies_app.User;
+import com.example.fortune_cookies_app.controller.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
